+++ src/main/java/net/frozenblock/freezeframe/networking/packet/CameraTakeScreenshotPacket.java
@@ -17,6 +17,7 @@
 
 package net.frozenblock.freezeframe.networking.packet;
 
+import java.util.OptionalInt;
 import net.fabricmc.fabric.api.networking.v1.ServerPlayNetworking;
 import net.frozenblock.freezeframe.FFConstants;
 import net.frozenblock.freezeframe.component.FilmFilter;
@@ -26,7 +27,6 @@ import net.minecraft.network.codec.ByteBufCodecs;
 import net.minecraft.network.codec.StreamCodec;
 import net.minecraft.network.protocol.common.custom.CustomPacketPayload;
 import net.minecraft.server.level.ServerPlayer;
-import java.util.OptionalInt;
 
 public record CameraTakeScreenshotPacket(OptionalInt entityId, boolean handheldCapture, float zoom, String fileName, FilmFilter filter) implements CustomPacketPayload {
 	public static final Type<CameraTakeScreenshotPacket> PACKET_TYPE = CustomPacketPayload.createType(FFConstants.safeString("camera_take_screenshot"));