+++ CHANGELOG.md
@@ -6,3 +6,4 @@ Put the changelog BELOW the dashes. ANYTHING ABOVE IS IGNORED.
   - Added the `FRAME_BACK` Frame Type.
 - Added separate textures for the Camera item, resolving an issue that locked mipmapping to 3 levels.
 - Fixed an issue that would cause an error about the `ContainerComponentManipulators` class to log, despite working as intended.
+- The Film Capacity Upgrade Recipe now supports a custom Source, Material, and Output item like other Vanilla transmute recipes.